Hopefully cheaper tests
Use a runtime xorshift* generater to create random memory access.
This should make the test less dependent on memory latency which should allow it to run faster.
Compared to using a pre generated list of random numbers, using a runtime generator
also increasing the memory access per-loop which could make the memory access test much harder
to interpret...
As a bonus, this should also works for store and not just load.